+++ src/org/db/dao/impl/ContactDAOImpl.java
@@ -15,20 +15,20 @@ import java.util.List;
 public class ContactDAOImpl implements ContactDAO {
     public List getAllContacts() throws SQLException {
         List contacts;
-        contacts = DAOHelpers.ExecuteInSession(session -> session.createCriteria(Contact.class).list(), false);
+        contacts = DAOHelpers.ExecuteInSession(session -> session.createCriteria(Contact.class).list());
         return contacts;
     }
 
     public Contact getContactById(Integer contactId) throws SQLException {
-        return DAOHelpers.ExecuteInSession(session -> (Contact) session.load(Contact.class, contactId), false);
+        return DAOHelpers.ExecuteInSession(session -> (Contact) session.get(Contact.class, contactId));
     }
 
     public int addContact(Contact w) throws SQLException {
-        return DAOHelpers.ExecuteInSession(session -> (int) session.save(w), true);
+        return DAOHelpers.ExecuteInSession(session -> (int) session.save(w));
     }
 
     public void updateContact(Contact w) throws SQLException {
-        DAOHelpers.ExecuteInSessionVoidRet(session -> session.update(w), true);
+        DAOHelpers.ExecuteInSessionVoidRet(session -> session.update(w));
     }
 
     public void deleteContact(Contact w) throws SQLException {
@@ -36,15 +36,17 @@ public class ContactDAOImpl implements ContactDAO {
             w.getCustomer().getContactSet().remove(w);
             session.save(w.getCustomer());
             session.delete(w);
-        }, true);
+        });
     }
 
     public void deleteContactById(int id) throws SQLException {
         DAOHelpers.ExecuteInSessionVoidRet(session -> {
-            Contact w = (Contact) session.load(Contact.class, id);
-            w.getCustomer().getContactSet().remove(w);
-            session.save(w.getCustomer());
-            session.delete(w);
-        }, true);
+            Contact w = (Contact) session.get(Contact.class, id);
+            if(w != null) {
+                w.getCustomer().getContactSet().remove(w);
+                session.save(w.getCustomer());
+                session.delete(w);
+            }
+        });
     }
 }
